Move from MD5 to SHA1
Take the SHA1 code out of Git.  Not only is it faster than the OpenSSL
version, but it is under a GPL-compatible license.

The only changes I made were to add in the #includes necessary for it
to compile (instead of the git-specific header), and to change an
'int' to 'unsigned' to eliminate a compiler warning.

The generated surefiles will have 'sha1' attributes instead of 'md5'
and will be slightly larger.