Add 'update' command
The update command reads the previous scan data, and uses it to
augment the tree read from the filesystem.  Files that have the same
path, inode and ctime will have the rest of their attributes copied
from the old file, avoiding the computation of the file hash for files
that haven't changed.